PR fixes
* Minor formatting changes
